pg_data/serializers.py
from rest_framework import serializers
from .models import Room, Tenant, Expense, TenantForm, Rent

class TenantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tenant
        fields = '__all__'

class RoomSerializer(serializers.ModelSerializer):

    class Meta:
        model = Room
        fields = '__all__'

class ExpenseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Expense
        fields = '__all__'

class TenantFormResponseSerializer(serializers.ModelSerializer):
    class Meta:
        model = TenantForm
        fields = '__all__'

class RentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Rent
        fields = '__all__'