requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
whitenoise
psycopg2-binary
